feat: implement grouped settings
