+++ src/views/footerView.jsx
@@ -21,10 +21,7 @@ import {
   MusicIcon,
   VerticalDivider,
   MessageModal,
-  ContextMenu,
-  ContextMenuItem,
   useHMSActions,
-  Settings,
   selectIsLocalScreenShared,
   selectIsLocalAudioEnabled,
   selectIsLocalVideoDisplayEnabled,
@@ -33,76 +30,12 @@ import {
   selectIsAllowedToPublish,
   selectIsLocalVideoPluginPresent,
   selectPermissions,
-  HamburgerMenuIcon,
-  SettingsIcon,
 } from "@100mslive/hms-video-react";
 import { useHistory, useParams } from "react-router-dom";
 import { HMSVirtualBackgroundPlugin } from "@100mslive/hms-virtual-background";
 import { AppContext } from "../store/AppContext";
 import { getRandomVirtualBackground } from "../common/utils";
-
-const SettingsView = () => {
-  const { setMaxTileCount, maxTileCount } = useContext(AppContext);
-  const [showMenu, setShowMenu] = useState(false);
-  const [showSettings, setShowSettings] = useState(false);
-
-  const onChange = count => {
-    setMaxTileCount(count);
-  };
-  return (
-    <Fragment>
-      <ContextMenu
-        menuOpen={showMenu}
-        onTrigger={value => {
-          console.log("onTrigger", value);
-          setShowMenu(value);
-        }}
-        classes={{
-          root: "static",
-          trigger: "bg-transparent-0",
-          menu: "mt-0 py-0",
-        }}
-        trigger={
-          <Button
-            iconOnly
-            variant="no-fill"
-            iconSize="md"
-            shape="rectangle"
-            active={showMenu}
-            onClick={() => {}}
-          >
-            <HamburgerMenuIcon />
-          </Button>
-        }
-        menuProps={{
-          anchorOrigin: {
-            vertical: "top",
-            horizontal: "center",
-          },
-          transformOrigin: {
-            vertical: "bottom",
-            horizontal: "center",
-          },
-        }}
-      >
-        <ContextMenuItem
-          icon={<SettingsIcon />}
-          label="Settings"
-          key="settings"
-          onClick={() => setShowSettings(true)}
-        />
-      </ContextMenu>
-      <Settings
-        className="hidden"
-        onTileCountChange={onChange}
-        maxTileCount={maxTileCount}
-        classes={{ sliderContainer: "hidden md:block" }}
-        showModal={showSettings}
-        onModalClose={() => setShowSettings(false)}
-      />
-    </Fragment>
-  );
-};
+import { MoreSettings } from "./components/MoreSettings";
 
 export const ConferenceFooter = ({ isChatOpen, toggleChat }) => {
   const isScreenShared = useHMSStore(selectIsLocalScreenShared);
@@ -302,7 +235,7 @@ export const ConferenceFooter = ({ isChatOpen, toggleChat }) => {
           <span key={4} className="mx-2 md:mx-3"></span>,
           <VerticalDivider key={5} />,
           <span key={6} className="mx-2 md:mx-3"></span>,
-          <SettingsView key={7} />,
+          <MoreSettings key={7} />,
         ]}
         rightComponents={[
           permissions?.endRoom ? (
